Refactoring of frontend to all be included in the /frontend folder
