--- my-react-app/src/Navbar.css
@@ -1,68 +1,0 @@
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: #12141b;
-  padding: 0 20px;
-  height: 60px; 
-  color: white;
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.nav-left,
-.nav-right a {
-  font-size: 1.5em; 
-  text-decoration: none; 
-  transition: color 0.3s;
-}
-
-.nav-right a:link,
-.nav-right a:visited,
-.nav-left a:visited,
-.nav-right a:active {
-  color: #ffffff; 
-}
-
-.nav-left:hover,
-.nav-right a:hover {
-  color: #535bf2; 
-}
-
-
-.nav-center {
-  display: flex;
-  align-items: center;
-  gap: 24px; 
-}
-
-.nav-center a {
-  text-decoration: none;
-  color: white;
-  font-size: 16px;
-  font-weight: bold;
-}
-
-.navbar svg {
-  font-size: 20px;
-  margin-right: 8px; 
-  cursor: pointer;
-}
-
-
-@media (max-width: 768px) {
-  .nav-center {
-    gap: 12px;
-  }
-
-  .nav-left,
-  .nav-right {
-    gap: 8px;
-  }
-
-  .navbar svg {
-    font-size: 18px;
-    margin-right: 6px;
-  }
-}
-
